+++ src/components/navigation-accordion/navigation-accordion.js
@@ -239,12 +239,7 @@ class NavigationAccordion extends React.PureComponent {
           <NavigationDropdown
             id={props.navigationDropdownId}
             currentPath={props.currentPath}
-            options={props.contents.firstLevelItems.map((option) => {
-              return {
-                label: option.title,
-                value: option.path
-              };
-            })}
+            dropdownOptions={props.contents.firstLevelItems}
             onChange={props.onDropdownChange}
           />
         </div>