+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Layouts
 
-- Update `NavigationDropdown` to use accessible menu.
-  - 🚨`dropdownOptions` prop for `NavigationDropdown` is now `options`.
+- Update `NavigationDropdown` to use accessible menu. [#267](https://github.com/mapbox/dr-ui/pull/267)
 
 ## 0.29.2